[algorithms][strategies] Move CS-specific code of envelope_segment from algorithms details to strategies.
This fixes circular dependencies.
