+++ include/boost/geometry/algorithms/detail/disjoint/segment_box.hpp
@@ -39,6 +39,9 @@
 
 #include <boost/geometry/geometries/box.hpp>
 
+// Temporary, for envelope_segment_impl
+#include <boost/geometry/strategies/spherical/envelope_segment.hpp>
+
 namespace boost { namespace geometry
 {
 
@@ -151,12 +154,15 @@ struct disjoint_segment_box_sphere_or_spheroid
 
         geometry::model::box<segment_point_type> box_seg;
 
-        geometry::detail::envelope::envelope_segment_impl<segment_cs_type>
-                ::template apply<geometry::radian>(lon1, lat1,
-                                                   lon2, lat2,
-                                                   box_seg,
-                                                   azimuth_strategy,
-                                                   alp1);
+        // TODO: Shouldn't CSTag be taken from the caller, not from the segment?
+        strategy::envelope::detail::envelope_segment_impl
+            <
+                segment_cs_type
+            >::template apply<geometry::radian>(lon1, lat1,
+                                                lon2, lat2,
+                                                box_seg,
+                                                azimuth_strategy,
+                                                alp1);
 
         if (disjoint_box_box(box, box_seg, disjoint_box_box_strategy))
         {
