+++ include/boost/geometry/algorithms/detail/envelope/segment.hpp
@@ -19,31 +19,10 @@
 #define BOOST_GEOMETRY_ALGORITHMS_DETAIL_ENVELOPE_SEGMENT_HPP
 
 #include <cstddef>
-#include <utility>
 
-#include <boost/core/ignore_unused.hpp>
-#include <boost/numeric/conversion/cast.hpp>
-
-#include <boost/geometry/core/assert.hpp>
-#include <boost/geometry/core/coordinate_system.hpp>
-#include <boost/geometry/core/coordinate_type.hpp>
-#include <boost/geometry/core/cs.hpp>
-#include <boost/geometry/core/point_type.hpp>
-#include <boost/geometry/core/radian_access.hpp>
 #include <boost/geometry/core/tags.hpp>
 
-#include <boost/geometry/util/math.hpp>
-
-#include <boost/geometry/geometries/helper_geometry.hpp>
-
-#include <boost/geometry/formulas/meridian_segment.hpp>
-#include <boost/geometry/formulas/vertex_latitude.hpp>
-
 #include <boost/geometry/algorithms/detail/assign_indexed_point.hpp>
-#include <boost/geometry/algorithms/detail/envelope/point.hpp>
-#include <boost/geometry/algorithms/detail/envelope/transform_units.hpp>
-#include <boost/geometry/algorithms/detail/expand/point.hpp>
-
 #include <boost/geometry/algorithms/dispatch/envelope.hpp>
 
 // For backward compatibility
@@ -58,384 +37,6 @@ namespace boost { namespace geometry
 namespace detail { namespace envelope
 {
 
-template <typename CalculationType, typename CS_Tag>
-struct envelope_segment_call_vertex_latitude
-{
-    template <typename T1, typename T2, typename Strategy>
-    static inline CalculationType apply(T1 const& lat1,
-                                        T2 const& alp1,
-                                        Strategy const& )
-    {
-        return geometry::formula::vertex_latitude<CalculationType, CS_Tag>
-            ::apply(lat1, alp1);
-    }
-};
-
-template <typename CalculationType>
-struct envelope_segment_call_vertex_latitude<CalculationType, geographic_tag>
-{
-    template <typename T1, typename T2, typename Strategy>
-    static inline CalculationType apply(T1 const& lat1,
-                                        T2 const& alp1,
-                                        Strategy const& strategy)
-    {
-        return geometry::formula::vertex_latitude<CalculationType, geographic_tag>
-            ::apply(lat1, alp1, strategy.model());
-    }
-};
-
-template <typename Units, typename CS_Tag>
-struct envelope_segment_convert_polar
-{
-    template <typename T>
-    static inline void pre(T & , T & ) {}
-
-    template <typename T>
-    static inline void post(T & , T & ) {}
-};
-
-template <typename Units>
-struct envelope_segment_convert_polar<Units, spherical_polar_tag>
-{
-    template <typename T>
-    static inline void pre(T & lat1, T & lat2)
-    {
-        lat1 = math::latitude_convert_ep<Units>(lat1);
-        lat2 = math::latitude_convert_ep<Units>(lat2);
-    }
-
-    template <typename T>
-    static inline void post(T & lat1, T & lat2)
-    {
-        lat1 = math::latitude_convert_ep<Units>(lat1);
-        lat2 = math::latitude_convert_ep<Units>(lat2);
-        std::swap(lat1, lat2);
-    }
-};
-
-template <typename CS_Tag>
-class envelope_segment_impl
-{
-private:
-
-    // degrees or radians
-    template <typename CalculationType>
-    static inline void swap(CalculationType& lon1,
-                            CalculationType& lat1,
-                            CalculationType& lon2,
-                            CalculationType& lat2)
-    {
-        std::swap(lon1, lon2);
-        std::swap(lat1, lat2);
-    }
-
-    // radians
-    template <typename CalculationType>
-    static inline bool contains_pi_half(CalculationType const& a1,
-                                        CalculationType const& a2)
-    {
-        // azimuths a1 and a2 are assumed to be in radians
-        BOOST_GEOMETRY_ASSERT(! math::equals(a1, a2));
-
-        static CalculationType const pi_half = math::half_pi<CalculationType>();
-
-        return (a1 < a2)
-                ? (a1 < pi_half && pi_half < a2)
-                : (a1 > pi_half && pi_half > a2);
-    }
-
-    // radians or degrees
-    template <typename Units, typename CoordinateType>
-    static inline bool crosses_antimeridian(CoordinateType const& lon1,
-                                            CoordinateType const& lon2)
-    {
-        typedef math::detail::constants_on_spheroid
-            <
-                CoordinateType, Units
-            > constants;
-
-        return math::abs(lon1 - lon2) > constants::half_period(); // > pi
-    }
-
-    // degrees or radians
-    template <typename Units, typename CalculationType, typename Strategy>
-    static inline void compute_box_corners(CalculationType& lon1,
-                                           CalculationType& lat1,
-                                           CalculationType& lon2,
-                                           CalculationType& lat2,
-                                           CalculationType a1,
-                                           CalculationType a2,
-                                           Strategy const& strategy)
-    {
-        // coordinates are assumed to be in radians
-        BOOST_GEOMETRY_ASSERT(lon1 <= lon2);
-        boost::ignore_unused(lon1, lon2);
-
-        CalculationType lat1_rad = math::as_radian<Units>(lat1);
-        CalculationType lat2_rad = math::as_radian<Units>(lat2);
-
-        if (math::equals(a1, a2))
-        {
-            // the segment must lie on the equator or is very short or is meridian
-            return;
-        }
-
-        if (lat1 > lat2)
-        {
-            std::swap(lat1, lat2);
-            std::swap(lat1_rad, lat2_rad);
-            std::swap(a1, a2);
-        }
-
-        if (contains_pi_half(a1, a2))
-        {
-            CalculationType p_max = envelope_segment_call_vertex_latitude
-                <CalculationType, CS_Tag>::apply(lat1_rad, a1, strategy);
-
-            CalculationType const mid_lat = lat1 + lat2;
-            if (mid_lat < 0)
-            {
-                // update using min latitude
-                CalculationType const lat_min_rad = -p_max;
-                CalculationType const lat_min
-                    = math::from_radian<Units>(lat_min_rad);
-
-                if (lat1 > lat_min)
-                {
-                    lat1 = lat_min;
-                }
-            }
-            else
-            {
-                // update using max latitude
-                CalculationType const lat_max_rad = p_max;
-                CalculationType const lat_max
-                    = math::from_radian<Units>(lat_max_rad);
-
-                if (lat2 < lat_max)
-                {
-                    lat2 = lat_max;
-                }
-            }
-        }
-    }
-
-    template <typename Units, typename CalculationType>
-    static inline void special_cases(CalculationType& lon1,
-                                     CalculationType& lat1,
-                                     CalculationType& lon2,
-                                     CalculationType& lat2)
-    {
-        typedef math::detail::constants_on_spheroid
-            <
-                CalculationType, Units
-            > constants;
-
-        bool is_pole1 = math::equals(math::abs(lat1), constants::max_latitude());
-        bool is_pole2 = math::equals(math::abs(lat2), constants::max_latitude());
-
-        if (is_pole1 && is_pole2)
-        {
-            // both points are poles; nothing more to do:
-            // longitudes are already normalized to 0
-            // but just in case
-            lon1 = 0;
-            lon2 = 0;
-        }
-        else if (is_pole1 && !is_pole2)
-        {
-            // first point is a pole, second point is not:
-            // make the longitude of the first point the same as that
-            // of the second point
-            lon1 = lon2;
-        }
-        else if (!is_pole1 && is_pole2)
-        {
-            // second point is a pole, first point is not:
-            // make the longitude of the second point the same as that
-            // of the first point
-            lon2 = lon1;
-        }
-
-        if (lon1 == lon2)
-        {
-            // segment lies on a meridian
-            if (lat1 > lat2)
-            {
-                std::swap(lat1, lat2);
-            }
-            return;
-        }
-
-        BOOST_GEOMETRY_ASSERT(!is_pole1 && !is_pole2);
-
-        if (lon1 > lon2)
-        {
-            swap(lon1, lat1, lon2, lat2);
-        }
-
-        if (crosses_antimeridian<Units>(lon1, lon2))
-        {
-            lon1 += constants::period();
-            swap(lon1, lat1, lon2, lat2);
-        }
-    }
-
-    template
-    <
-        typename Units,
-        typename CalculationType,
-        typename Box
-    >
-    static inline void create_box(CalculationType lon1,
-                                  CalculationType lat1,
-                                  CalculationType lon2,
-                                  CalculationType lat2,
-                                  Box& mbr)
-    {
-        typedef typename coordinate_type<Box>::type box_coordinate_type;
-
-        typedef typename helper_geometry
-            <
-                Box, box_coordinate_type, Units
-            >::type helper_box_type;
-
-        helper_box_type helper_mbr;
-
-        geometry::set
-            <
-                min_corner, 0
-            >(helper_mbr, boost::numeric_cast<box_coordinate_type>(lon1));
-
-        geometry::set
-            <
-                min_corner, 1
-            >(helper_mbr, boost::numeric_cast<box_coordinate_type>(lat1));
-
-        geometry::set
-            <
-                max_corner, 0
-            >(helper_mbr, boost::numeric_cast<box_coordinate_type>(lon2));
-
-        geometry::set
-            <
-                max_corner, 1
-            >(helper_mbr, boost::numeric_cast<box_coordinate_type>(lat2));
-
-        transform_units(helper_mbr, mbr);
-    }
-
-
-    template <typename Units, typename CalculationType, typename Strategy>
-    static inline void apply(CalculationType& lon1,
-                             CalculationType& lat1,
-                             CalculationType& lon2,
-                             CalculationType& lat2,
-                             Strategy const& strategy)
-    {
-        special_cases<Units>(lon1, lat1, lon2, lat2);
-
-        CalculationType lon1_rad = math::as_radian<Units>(lon1);
-        CalculationType lat1_rad = math::as_radian<Units>(lat1);
-        CalculationType lon2_rad = math::as_radian<Units>(lon2);
-        CalculationType lat2_rad = math::as_radian<Units>(lat2);
-        CalculationType alp1, alp2;
-        strategy.apply(lon1_rad, lat1_rad, lon2_rad, lat2_rad, alp1, alp2);
-
-        compute_box_corners<Units>(lon1, lat1, lon2, lat2, alp1, alp2, strategy);
-    }
-
-    template <typename Units, typename CalculationType, typename Strategy>
-    static inline void apply(CalculationType& lon1,
-                             CalculationType& lat1,
-                             CalculationType& lon2,
-                             CalculationType& lat2,
-                             Strategy const& strategy,
-                             CalculationType alp1)
-    {
-        special_cases<Units>(lon1, lat1, lon2, lat2);
-
-        CalculationType lon1_rad = math::as_radian<Units>(lon1);
-        CalculationType lat1_rad = math::as_radian<Units>(lat1);
-        CalculationType lon2_rad = math::as_radian<Units>(lon2);
-        CalculationType lat2_rad = math::as_radian<Units>(lat2);
-        CalculationType alp2;
-        strategy.apply_reverse(lon1_rad, lat1_rad, lon2_rad, lat2_rad, alp2);
-
-        compute_box_corners<Units>(lon1, lat1, lon2, lat2, alp1, alp2, strategy);
-    }
-
-public:
-    template
-    <
-        typename Units,
-        typename CalculationType,
-        typename Box,
-        typename Strategy
-    >
-    static inline void apply(CalculationType lon1,
-                             CalculationType lat1,
-                             CalculationType lon2,
-                             CalculationType lat2,
-                             Box& mbr,
-                             Strategy const& strategy)
-    {
-        typedef envelope_segment_convert_polar<Units, typename cs_tag<Box>::type> convert_polar;
-
-        convert_polar::pre(lat1, lat2);
-
-        apply<Units>(lon1, lat1, lon2, lat2, strategy);
-
-        convert_polar::post(lat1, lat2);
-
-        create_box<Units>(lon1, lat1, lon2, lat2, mbr);
-    }
-
-    template
-    <
-        typename Units,
-        typename CalculationType,
-        typename Box,
-        typename Strategy
-    >
-    static inline void apply(CalculationType lon1,
-                             CalculationType lat1,
-                             CalculationType lon2,
-                             CalculationType lat2,
-                             Box& mbr,
-                             Strategy const& strategy,
-                             CalculationType alp1)
-    {
-        typedef envelope_segment_convert_polar<Units, typename cs_tag<Box>::type> convert_polar;
-
-        convert_polar::pre(lat1, lat2);
-
-        apply<Units>(lon1, lat1, lon2, lat2, strategy, alp1);
-
-        convert_polar::post(lat1, lat2);
-
-        create_box<Units>(lon1, lat1, lon2, lat2, mbr);
-    }
-};
-
-template <std::size_t Dimension, std::size_t DimensionCount>
-struct envelope_one_segment
-{
-    template<typename Point, typename Box>
-    static inline void apply(Point const& p1,
-                             Point const& p2,
-                             Box& mbr)
-    {
-        envelope_one_point<Dimension, DimensionCount>::apply(p1, mbr);
-        strategy::expand::detail::point_loop
-            <
-                Dimension,
-                DimensionCount
-            >::apply(mbr, p2);
-    }
-};
-
-
 template <std::size_t DimensionCount>
 struct envelope_segment
 {
@@ -445,12 +46,7 @@ struct envelope_segment
                              Box& mbr,
                              Strategy const& strategy)
     {
-        // first compute the envelope range for the first two coordinates
         strategy.apply(p1, p2, mbr);
-
-        // now compute the envelope range for coordinates of
-        // dimension 2 and higher
-        envelope_one_segment<2, DimensionCount>::apply(p1, p2, mbr);
     }
 
     template <typename Segment, typename Box, typename Strategy>
