+++ include/boost/geometry/strategies/geographic/envelope_segment.hpp
@@ -12,14 +12,14 @@
 #define BOOST_GEOMETRY_STRATEGIES_GEOGRAPHIC_ENVELOPE_SEGMENT_HPP
 
 
-#include <boost/geometry/algorithms/detail/envelope/segment.hpp>
-#include <boost/geometry/algorithms/detail/normalize.hpp>
-
 #include <boost/geometry/srs/spheroid.hpp>
 
+#include <boost/geometry/strategies/cartesian/envelope_segment.hpp>
 #include <boost/geometry/strategies/envelope.hpp>
 #include <boost/geometry/strategies/geographic/azimuth.hpp>
 #include <boost/geometry/strategies/geographic/parameters.hpp>
+#include <boost/geometry/strategies/normalize.hpp>
+#include <boost/geometry/strategies/spherical/envelope_segment.hpp>
 #include <boost/geometry/strategies/spherical/expand_box.hpp>
 
 namespace boost { namespace geometry
@@ -53,12 +53,11 @@ public:
         return box_expand_strategy_type();
     }
 
-    template <typename Point1, typename Point2, typename Box>
-    inline void apply(Point1 const& point1, Point2 const& point2, Box& box) const
+    template <typename Point, typename Box>
+    inline void apply(Point const& point1, Point const& point2, Box& box) const
     {
-        Point1 p1_normalized;
+        Point p1_normalized, p2_normalized;
         strategy::normalize::spherical_point::apply(point1, p1_normalized);
-        Point2 p2_normalized;
         strategy::normalize::spherical_point::apply(point2, p2_normalized);
 
         geometry::strategy::azimuth::geographic
@@ -70,10 +69,11 @@ public:
 
         typedef typename geometry::detail::cs_angular_units
             <
-                Point1
+                Point
             >::type units_type;
 
-        detail::envelope::envelope_segment_impl
+        // first compute the envelope range for the first two coordinates
+        strategy::envelope::detail::envelope_segment_impl
             <
                 geographic_tag
             >::template apply<units_type>(geometry::get<0>(p1_normalized),
@@ -83,6 +83,12 @@ public:
                                           box,
                                           azimuth_geographic);
 
+        // now compute the envelope range for coordinates of
+        // dimension 2 and higher
+        strategy::envelope::detail::envelope_one_segment
+            <
+                2, dimension<Point>::value
+            >::apply(point1, point2, box);
     }
 
 private:
